refactor(web): extract dereg shared constants
