+++ apps/web/src/app/(dashboard)/cars/deregistrations/_components/category-chart.tsx
@@ -8,66 +8,18 @@ import {
   ChartTooltip,
   ChartTooltipContent,
 } from "@sgcarstrends/ui/components/chart";
+import {
+  type CategoryWithPercentage,
+  toPercentageDistribution,
+} from "@web/app/(dashboard)/cars/deregistrations/_components/constants";
 import { deregistrationsSearchParams } from "@web/app/(dashboard)/cars/deregistrations/search-params";
 import Typography from "@web/components/typography";
-import {
-  chartColourPalette,
-  formatNumber,
-  formatPercentage,
-} from "@web/utils/charts";
+import { formatNumber, formatPercentage } from "@web/utils/charts";
 import { formatDateToMonthYear } from "@web/utils/format-date-to-month-year";
 import { useQueryStates } from "nuqs";
 import type React from "react";
 import { useMemo } from "react";
 import { Bar, BarChart, CartesianGrid, Cell, XAxis, YAxis } from "recharts";
-
-// Category colours
-const DEREGISTRATION_CATEGORIES = [
-  "Category A",
-  "Category B",
-  "Category C",
-  "Category D",
-  "Vehicles Exempted From VQS",
-  "Taxis",
-] as const;
-
-type DeregistrationCategory = (typeof DEREGISTRATION_CATEGORIES)[number];
-
-const DEREGISTRATION_CATEGORY_COLOURS: Record<DeregistrationCategory, string> =
-  {
-    "Category A": chartColourPalette[0],
-    "Category B": chartColourPalette[1],
-    "Category C": chartColourPalette[2],
-    "Category D": chartColourPalette[3],
-    "Vehicles Exempted From VQS": chartColourPalette[4],
-    Taxis: chartColourPalette[5],
-  };
-
-const getCategoryColour = (category: string): string => {
-  if (category in DEREGISTRATION_CATEGORY_COLOURS) {
-    return DEREGISTRATION_CATEGORY_COLOURS[category as DeregistrationCategory];
-  }
-  return chartColourPalette[0];
-};
-
-interface CategoryWithPercentage {
-  category: string;
-  total: number;
-  percentage: number;
-  colour: string;
-}
-
-const toPercentageDistribution = (
-  data: { category: string; total: number }[],
-): CategoryWithPercentage[] => {
-  const grandTotal = data.reduce((sum, item) => sum + item.total, 0);
-
-  return data.map((item) => ({
-    ...item,
-    percentage: grandTotal > 0 ? (item.total / grandTotal) * 100 : 0,
-    colour: getCategoryColour(item.category),
-  }));
-};
 
 interface Props {
   data: SelectDeregistration[];
